Add CI, build system, documentation, and DLL export macros
- Add Drone CI configuration (.drone.star, .drone/*)
- Expand GitHub Actions workflow with matrix builds
- Add B2/Jamfile build support alongside CMake
- Add Antora documentation scaffolding
- Add BOOST_BURL_DECL macro for DLL symbol export
- Use BOOST_BURL_DECL on session, cookie_jar, error_category,
  http_error, http_basic_auth, http_digest_auth, http_bearer_auth
- Reorganize test files into test/unit/ directory
- Add cmake_test for installation verification
