+++ doc/build_antora.sh
@@ -1,0 +1,28 @@
+#
+# Copyright (c) 2023 Alan de Freitas (alandefreitas@gmail.com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+# Official repository: https://github.com/cppalliance/burl
+#
+
+set -xe
+
+if [ $# -eq 0 ]
+  then
+    echo "No playbook supplied, using default playbook"
+    PLAYBOOK="local-playbook.yml"
+  else
+    PLAYBOOK=$1
+fi
+
+echo "Building documentation with Antora..."
+echo "Installing npm dependencies..."
+npm install
+
+echo "Building docs in custom dir..."
+PATH="$(pwd)/node_modules/.bin:${PATH}"
+export PATH
+npx antora --clean --fetch "$PLAYBOOK"
+echo "Done"
