+++ doc/package.json
@@ -1,0 +1,17 @@
+{
+  "devDependencies": {
+    "@antora/cli": "3.1.14",
+    "@antora/site-generator": "3.1.14",
+    "antora": "3.1.14"
+  },
+  "dependencies": {
+    "@antora/expand-path-helper": "^3.0.0",
+    "@antora/lunr-extension": "^1.0.0-alpha.12",
+    "@asciidoctor/tabs": "^1.0.0-beta.6",
+    "@cppalliance/antora-cpp-reference-extension": "^0.1.0",
+    "@cppalliance/antora-cpp-tagfiles-extension": "^0.1.0",
+    "@cppalliance/antora-downloads-extension": "^0.0.2",
+    "@cppalliance/asciidoctor-boost-links": "^0.0.2",
+    "@sntke/antora-mermaid-extension": "^0.0.8"
+  }
+}
