+++ example/Jamfile
@@ -1,0 +1,32 @@
+#
+# Copyright (c) 2025 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+# Official repository: https://github.com/cppalliance/burl
+#
+
+using openssl ;
+import ac ;
+
+lib advapi32 ;
+lib crypt32 ;
+lib gdi32 ;
+lib user32 ;
+
+project
+    : requirements
+      <library>/boost/burl//boost_burl
+      [ ac.check-library /openssl//ssl : <library>/openssl//ssl/<link>shared : <build>no ]
+      [ ac.check-library /openssl//crypto : <library>/openssl//crypto/<link>shared : <build>no ]
+      <target-os>windows:<library>advapi32
+      <target-os>windows:<library>crypt32
+      <target-os>windows:<library>gdi32
+      <target-os>windows:<library>user32
+      <include>.
+    ;
+
+exe usage :
+    [ glob *.cpp ]
+    ;
