+++ include/boost/burl.hpp
@@ -1,0 +1,22 @@
+//
+// Copyright (c) 2025 Vinnie Falco (vinnie dot falco at gmail dot com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+// Official repository: https://github.com/cppalliance/burl
+//
+
+#ifndef BOOST_BURL_HPP
+#define BOOST_BURL_HPP
+
+#include <boost/burl/auth.hpp>
+#include <boost/burl/body_tags.hpp>
+#include <boost/burl/cookies.hpp>
+#include <boost/burl/error.hpp>
+#include <boost/burl/fwd.hpp>
+#include <boost/burl/options.hpp>
+#include <boost/burl/response.hpp>
+#include <boost/burl/session.hpp>
+
+#endif
