+++ test/cmake_test/CMakeLists.txt
@@ -1,0 +1,27 @@
+#
+# Copyright (c) 2025 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0.
+# https://www.boost.org/LICENSE_1_0.txt
+#
+
+cmake_minimum_required(VERSION 3.5...3.16)
+
+project(cmake_subdir_test LANGUAGES CXX)
+set(__ignore__ ${CMAKE_C_COMPILER})
+set(__ignore__ ${CMAKE_C_FLAGS})
+
+if(BOOST_CI_INSTALL_TEST)
+  find_package(Boost CONFIG REQUIRED COMPONENTS burl)
+else()
+  set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
+  set(BOOST_INCLUDE_LIBRARIES burl)
+  add_subdirectory(../../../.. boostorg/boost)
+endif()
+
+add_executable(main main.cpp)
+target_link_libraries(main Boost::burl)
+
+enable_testing()
+add_test(NAME main COMMAND main)
+add_custom_target(check COMMAND ${CMAKE_CTEST_COMMAND} --output-on-failure -C $<CONFIG>)
