+++ test/cmake_test/main.cpp
@@ -1,0 +1,7 @@
+#include <boost/burl.hpp>
+
+int main() {
+  std::error_code ec = boost::burl::error::success;
+  if (ec)
+    throw;
+}
