+++ test/unit/CMakeLists.txt
@@ -1,0 +1,33 @@
+#
+# Copyright (c) 2025 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+# Official repository: https://github.com/cppalliance/burl
+#
+
+if(NOT TARGET boost_capy_test_suite)
+    add_subdirectory(../../../capy/extra/test_suite test_suite)
+endif()
+
+file(GLOB_RECURSE PFILES CONFIGURE_DEPENDS *.cpp *.hpp)
+list(APPEND PFILES
+    CMakeLists.txt
+    Jamfile)
+
+source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} PREFIX "" FILES ${PFILES})
+
+add_executable(boost_burl_tests ${PFILES})
+target_link_libraries(
+    boost_burl_tests PRIVATE
+    boost_capy_test_suite_main
+    Boost::burl)
+
+target_include_directories(boost_burl_tests PRIVATE . ../../)
+
+# Register individual tests with CTest
+boost_capy_test_suite_discover_tests(boost_burl_tests)
+
+# Add the main test target to Boost's test suite
+add_dependencies(tests boost_burl_tests)
