+++ test/unit/Jamfile
@@ -1,0 +1,29 @@
+#
+# Copyright (c) 2025 Vinnie Falco (vinnie dot falco at gmail dot com)
+#
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#
+# Official repository: https://github.com/cppalliance/burl
+#
+
+import testing ;
+
+project
+    : requirements
+      $(c20-requires)
+      <library>/boost/burl//boost_burl
+      <source>../../../capy/extra/test_suite/test_main.cpp
+      <source>../../../capy/extra/test_suite/test_suite.cpp
+      <include>.
+      <include>../..
+      <include>../../../capy/extra/test_suite
+      <warnings>extra
+      # <warnings-as-errors>on
+      <undefined-sanitizer>norecover:<link>static
+    ;
+
+for local f in [ glob *.cpp ]
+{
+    run $(f) ;
+}
